Fix compile warning

+++ src/vw/tools/make_ba_test_data.cc
@@ -1001,7 +1001,7 @@ void write_control_network(boost::shared_ptr<ControlNetwork> cnet, int num_camer
     ControlPoint cp = *cnet_iter;
     Vector3 pos = cp.position();
     cnetos << pos[0] << "\t" << pos[1] << "\t" << pos[2] << "\t";
-    for (size_t i = 0; i < num_cameras; i++) {
+    for (int i = 0; i < num_cameras; i++) {
       bool match = false;
       for (size_t j = 0; j < cp.size(); j++) {
         ControlMeasure cm = cp[j];
